Extract quizpage logic into quiz folder;types, and utilities in separate modules
