+++ src/pages/quiz/types.ts
@@ -1,0 +1,4 @@
+export interface LayoutChoice { key: string; text: string }
+export interface LayoutItem { question: string; questionNumber: string | number; choices: LayoutChoice[] }
+
+
